Fix Windows build regressions.

+++ libtiledbsoma/src/reindexer/reindexer.cc
@@ -32,7 +32,6 @@
 
 #include "reindexer.h"
 #include <thread_pool/thread_pool.h>
-#include <unistd.h>
 #include <thread>
 #include "khash.h"
 #include "soma/enums.h"
@@ -78,14 +77,14 @@ void IntIndexer::map_locations(const int64_t* keys, size_t size) {
         fmt::format("[Re-indexer] Thread pool started and hash table created"));
 }
 
-void IntIndexer::lookup(const int64_t* keys, int64_t* results, int size) {
+void IntIndexer::lookup(const int64_t* keys, int64_t* results, size_t size) {
     if (size == 0) {
         return;
     }
     // Single thread checks
     if (context_ == nullptr || context_->thread_pool() == nullptr ||
         context_->thread_pool()->concurrency_level() == 1) {
-        for (int i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             auto k = kh_get(m64, hash_, keys[i]);
             if (k == kh_end(hash_)) {
                 // According to pandas behavior
@@ -109,10 +108,10 @@ void IntIndexer::lookup(const int64_t* keys, int64_t* results, int size) {
         thread_chunk_size = 1;
     }
 
-    for (size_t i = 0; i < size_t(size); i += thread_chunk_size) {
+    for (size_t i = 0; i < size; i += thread_chunk_size) {
         size_t start = i;
         size_t end = i + thread_chunk_size;
-        if (end > size_t(size)) {
+        if (end > size) {
             end = size;
         }
         LOG_DEBUG(fmt::format(
